added sorted submission report to service
